feat: create resource w/ relationship key immediately
This does affect resource relationships which are of type `BelongsTo`
and `HasOne` and therefore have a foreign key field within the same table
(`relationshipId`). This change was necessary as it is common to have
non nullable foreign keys and the previously used `setRelationship`
need an extra query to set the foreign key value.