feat: add scopes to extend queries
This is a powerful new feature, as it allows you to manipulate queries sent
to the database.

In this example, we have added a new scope `byEmail` to the `User` model
which adds a `where` clause that filters the result by a given email
address as query parameter.

```js
// models/user.js
const User = db.define('User', {
  id: {
    type: Sequelize.UUID,
    primaryKey: true
  },
  email: {
    type: Sequelize.STRING,
  }
}, {
  scopes: {
    byEmail: function(req) {
      return {
        where: {email: req.query.email}
      }
    }
  }
})
```

```js
// routes/users.js
api.rest({
  model: User,
  scopes: ['defaultScope', 'byEmail']
})
```

Definition of a new scope must follow the following pattern:

```js
{
  name: function(req, res) {
    return {…}
  }
}
```